fix: tipe data response in route group entity and model

+++ lib/data/models/route_group_model.dart
@@ -16,9 +16,9 @@ class RouteGroupModel extends RouteGroup{
     sellCost: sellCost
   );
 
-  final int minutes;
+  final double minutes;
   final String name;
-  final int sellCost;
+  final double sellCost;
 
   factory RouteGroupModel.fromJson(Map<String, dynamic> json) => RouteGroupModel(
       minutes: json["minutes"],
